// client/src/store/action-types/election.action.types.ts
export enum electionActionTypes {
  ELECTION_LIST_REQUEST = 'ELECTION_LIST_REQUEST',
  ELECTION_LIST_SUCCESS = 'ELECTION_LIST_SUCCESS',
  ELECTION_LIST_FAIL = 'ELECTION_LIST_FAIL',
  ELECTION_DETAILS_REQUEST = 'ELECTION_DETAILS_REQUEST',
  ELECTION_DETAILS_SUCCESS = 'ELECTION_DETAILS_SUCCESS',
  ELECTION_DETAILS_FAIL = 'ELECTION_DETAILS_FAIL',
  ELECTION_CREATE_REQUEST = 'ELECTION_CREATE_REQUEST',
  ELECTION_CREATE_SUCCESS = 'ELECTION_CREATE_SUCCESS',
  ELECTION_CREATE_FAIL = 'ELECTION_CREATE_FAIL',
  ELECTION_CREATE_RESET = 'ELECTION_CREATE_RESET',
  ELECTION_END_REQUEST = 'ELECTION_END_REQUEST',
  ELECTION_END_SUCCESS = 'ELECTION_END_SUCCESS',
  ELECTION_END_FAIL = 'ELECTION_END_FAIL',
}